grammar: Fix built-in type list
Some of the types listed as “primitve types” are indeed primitive, but
not built-in as far as the language standard goes. Moreover, a few are
missing, so this commit removes the common definitions, such as
{u,}intN_t and friends and adds the fancy C99 _Bool, _Complex and
_Imaginary.

Since the change broke existing tests, this additionally updates the
corpus. New checks are added to see if the common type definitions are
not mistaken for built-in types.

The downside of this all would be that those common definitions cease to
be highlighted.

Closes https://github.com/tree-sitter/tree-sitter-c/issues/52