fix: allow preprocessor directives in an `enum_specifier`
Preprocessor directives inside of an `enum_specifier` variant list cause
ERROR nodes to appear in the syntax tree. Both the `struct_specifier`
and the `union_specifier` allow the preprocessor directives to exist to
selectively enable or disable fields in the `struct` or `union`. It is
fairly common in C to also selectively add variants to an `enum` in the
same manner.

This modifies the grammar for the `enumerator_list` so that the
preprocessor directives are allowed. My initial approach was to just
make the commas optional at the end of the `enumerator` and, while that
did allow a simple `_enumerator_list_item` in much the same style as the
`_field_declaration_list_item`, it was a change in behavior since it did
not require a comma in between every `enum` variant. By splitting the
`_enumerator_list_item` into a version with a comma required and an
optional `_enumerator_list_item_end` that does NOT require a comma, this
keeps the same behvior of requiring the comma separated list of variants
without requiring a comma at the end of the list of variants.
Additionally, it does not require a comma at the end of the preprocessor
directives, which would not be correct either. The conflict additions
were required to get the grammar to compile with the two different
versions of `enumerator_list_item`.

Additionally, the modifications to the grammar introduced some ambiguity
in the precedence of the `concatenated_string`, so right precedence was
remove that ambiguity.

Fixes #181